+++ queue_services/entity-filer/src/entity_filer/filing_processors/change_of_registration.py
@@ -29,7 +29,7 @@ from entity_filer.filing_processors.filing_components import (
 )
 from entity_filer.filing_processors.filing_components.parties import get_or_create_party, merge_all_parties
 from entity_filer.filing_processors.registration import get_partnership_name
-from entity_filer.filing_processors.filing_components.partner import update_partner_change, update_proprietor_change
+from entity_filer.filing_processors.filing_components.alternate_name import update_partner_change, update_proprietor_change
 
 
 def  process(
